Sửa lại các kiểu dữ liệu
Sửa lại các kiểu dữ liệu trong migrations

+++ database/migrations/2019_10_29_143826_create__cau_hinh_tro_giups_table.php
@@ -15,9 +15,9 @@ class CreateCauHinhTroGiupsTable extends Migration
     {
         Schema::create('cau_hinh_tro_giup', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('loai_tro_giup');
-            $table->bigInteger('thu_tu');
-            $table->bigInteger('credit');
+            $table->Integer('loai_tro_giup');
+            $table->Integer('thu_tu');
+            $table->Integer('credit');
             $table->timestamps();
         });
     }
